requirements.txt
pandas
numpy
matplotlib
scikit-learn
pulp
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
websockets
httpx
apscheduler
pydantic-settings
python-multipart